+++ src/index.js
@@ -1,2 +1,2 @@
 export { CREATE, DELETE, INCREMENT, SET } from './utils/mutation.js'
-export { withIota } from './utils/with-iota.js'
+export { withIota } from './utils/withIota.js'